+++ docs/.vuepress/config.js
@@ -1,4 +1,10 @@
 module.exports = {
   title: 'JS-Tips',
-  description: '前端基础技巧',
+  description: '前端基础',
+  themeConfig: {
+    nav: [
+      { text: '主页', link: '/' },
+      { text: 'Github', link: 'https://github.com/JiangWeixian/JS-Tips/' },
+    ]
+  }
 }