+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## How to build vuepress blog
 
-* []
+* [build-vuepress-blog](https://juejin.im/post/5addb90af265da0b7f442935)